quote_walls/pyproject.toml
[project]
name = "quote-walls"
version = "0.1.0"
description = "A script to generate and set wallpapers from quotes I like"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=10.1.0",
]
